trying to draw chart

+++ stocks/views.py
@@ -1,10 +1,31 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Stock
-from .forms import StockForm  # Create a form for adding/editing stock data
-
+from .forms import StockForm  
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.shortcuts import render
+from .models import Stock
 def home(request):
+    
     stocks = Stock.objects.all()
-    return render(request, 'stocks/home.html', {'stocks': stocks})
+
+    trade_codes = Stock.objects.values_list('trade_code', flat=True).distinct()
+    stocks_json = serialize_stock_data(stocks)
+
+    
+    paginator = Paginator(stocks, 10)  
+    page = request.GET.get('page')
+
+    try:
+        stocks = paginator.page(page)
+    except PageNotAnInteger:
+        
+        stocks = paginator.page(1)
+    except EmptyPage:
+        
+        stocks = paginator.page(paginator.num_pages)
+
+    # return render(request, 'stocks/home.html', {'stocks': stocks})  
+    return render(request, 'stocks/home.html', {'stocks': stocks, 'trade_codes': trade_codes, 'stocks_json': stocks_json})
 
 def stock_detail(request, pk):
     stock = get_object_or_404(Stock, pk=pk)
@@ -15,10 +36,26 @@ def stock_create(request):
         form = StockForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('home')
+            return render(request, 'success.html')
     else:
         form = StockForm()
     return render(request, 'stocks/stock_form.html', {'form': form})
+
+# def stock_edit(request, pk):
+#     stock = get_object_or_404(Stock, pk=pk)
+#     if request.method == 'POST':
+#         form = StockForm(request.POST, instance=stock)
+#         if form.is_valid():
+#             form.save()
+#             return redirect('home')
+#     else:
+#         form = StockForm(instance=stock)
+#     return render(request, 'stocks/stock_form.html', {'form': form})
+
+# def stock_delete(request, pk):
+#     stock = get_object_or_404(Stock, pk=pk)
+#     stock.delete()
+#     return redirect('home')
 
 def stock_edit(request, pk):
     stock = get_object_or_404(Stock, pk=pk)
@@ -33,5 +70,25 @@ def stock_edit(request, pk):
 
 def stock_delete(request, pk):
     stock = get_object_or_404(Stock, pk=pk)
-    stock.delete()
-    return redirect('home')
+    if request.method == 'POST':
+        stock.delete()
+        return redirect('home')
+    return render(request, 'stocks/stock_confirm_delete.html', {'stock': stock})
+
+
+
+def serialize_stock_data(stocks):
+    serialized_data = []
+    for stock in stocks:
+        serialized_data.append({
+            'date': stock.date.strftime('%Y-%m-%d'),
+            'trade_code': stock.trade_code,
+            'high': stock.high,
+            'low': stock.low,
+            'open': stock.open,
+            'close': stock.close,
+            'volume': stock.volume,
+        })
+    return serialized_data
+
+
